test(proto): Create tests::token module
Moves the existing `stateless_retry` test into that module.
